Pbft base number changed to float64 transmission

+++ pbftserver/consensus/pbft_impl.go
@@ -199,7 +199,7 @@ func (state *State) PrePrepare(prePrepareMsg *PrePrepareMsg) (*VoteMsg, error) {
 	}, nil
 }
 
-func (state *State) Prepare(prepareMsg *VoteMsg, f int) (*VoteMsg, error) {
+func (state *State) Prepare(prepareMsg *VoteMsg, f float64) (*VoteMsg, error) {
 	//lock.PSLog("Prepare in")
 	if !state.verifyMsg(prepareMsg.ViewID, prepareMsg.SequenceID, prepareMsg.Digest) {
 		return nil, errors.New("prepare message is corrupted")
@@ -230,7 +230,7 @@ func (state *State) Prepare(prepareMsg *VoteMsg, f int) (*VoteMsg, error) {
 	return nil, nil
 }
 
-func (state *State) Commit(commitMsg *VoteMsg, f int) (*ReplyMsg, *RequestMsg, error) {
+func (state *State) Commit(commitMsg *VoteMsg, f float64) (*ReplyMsg, *RequestMsg, error) {
 
 	if !state.verifyMsg(commitMsg.ViewID, commitMsg.SequenceID, commitMsg.Digest) {
 		return nil, nil, errors.New("commit message is corrupted")
@@ -285,30 +285,30 @@ func (state *State) verifyMsg(viewID int64, sequenceID int64, digestGot string) 
 	return true
 }
 
-func (state *State) prepared(f int) bool {
+func (state *State) prepared(f float64) bool {
 	if state.MsgLogs.ReqMsg == nil {
 		return false
 	}
-	if state.MsgLogs.GetPrepareCount() < 2*f {
+	if state.MsgLogs.GetPrepareCount() < int(2*f) {
 		return false
 	}
 
 	return true
 }
 
-func (state *State) committed(f int) bool {
+func (state *State) committed(f float64) bool {
 	lock.PSLog("committed in")
 	if !state.prepared(f) {
 		return false
 	}
 	lock.PSLog("committed prepared")
-	if state.MsgLogs.GetCommitCount() <= 2*f {
+	if state.MsgLogs.GetCommitCount() <= int(2*f) {
 		return false
 	}
 	lock.PSLog("committed len(state.MsgLogs.CommitMsgs) >= 2*f")
 	passCount := state.MsgLogs.GetCommitPassCount()
 	//lock.PSLog("committed", fmt.Sprintf("%+v", state.MsgLogs.CommitMsgs), passCount)
-	return passCount > 2*f
+	return passCount > int(2*f)
 }
 
 func digest(object interface{}) (string, error) {
